Use std::unique_lock and std::condition_variable  for performance improv
Thanks to the ability to make the threads wait when there is no tasks for them, the constant mutex lock/unlock and checks that performs each iteration when using the std::lock_guard are now gone. One thread will be notified when a task is added to the queue and, if the requirements are met, it will lock the mutex and wake up the thread, so the task con be consumed or executed.